FIX Correct form field names that have been namespaced
The Silverstripe upgrader tool looks for strings that match classnames
and replaces them with the ::class shorthand, as this is a common pattern
in Silverstripe (with ORM queries). However in some cases, commonly form
fields, the string is a legitimate field name that just so happens to match
an un-namespaced classname - and should not be updated.