FIX update Bulk Archive Handler
The interface has changed heavily with GridFieldBulkEditingTools
and Silverstripe CMS 4. This update needs to reflect this.
Although the component is confusingly named as there is an archive
handler bundled with the bulk editing tools package, it works for
Versioned records, where these FAQ records are not versioned, but
instead have a checkbox to indicate their versioned status. This
functionality was kept in order to ease upgrade problems.