FIX Admin search/filtering fields not applying
When administering an FAQ section, using the supplied ModelAdmin
header search field would see some of the fields correctly apply
to the filtered results (or have no affect at all if they're the
only filed filled). This is due to API changes between Silverstripe
CMS 3 and Silverstripe CMS 4 - more is taken care of in a cleaner
manner, so in fact we need less complexity to achieve the result.