FIX bad query when filtering FAQSearch Admin
In older versions of Silverstripe it was reasonable to both presume
table names, and also use the `where` filter when `IS NULL` or the
NOT equivalent of it was desired. Silverstripe CMS 4 however has
decent support for the `null` type in a filter, and table names
can be aliased by the system - which resulted in the filter changed
in this commit causing a 500 server error due to a bad SQL query.