FIX Obsolete class calls with Silverstripe 4 compatible
Object is a reserved word in PHP 7+, and as such is not referenced
in Silverstripe CMS 3.7 and above. The calls here were not caught
by the upgrader tool, and have escaped detection until testing
revealed a 500 in the CMS when attempting to view "viewed articles"
from a FAQSearch admin detail form. Correcting these mistakes has
resolved the issue.